+++ index.ts
@@ -1,2 +1,1 @@
-// export * from './src/structure/Discord';
 export * as Discord from "./src";