+++ src/builder/index.ts
@@ -26,11 +26,23 @@ export {
 } from './UserCommandBuilder'
 
 export {
-    AccessoryBuilder as Accessory,
-    ButtonBuilder as Button,
     ComponentsBuilder as Components,
-    ActionRowBuilder as ActionRow,
-    ContainerBuilder as Container,
     type MessageComponentsBuilder as MessageComponents,
     type MessageContainerComponentsBuilder as MessageContainerComponents
 } from './MessageComponentsBuilder'
+
+export {
+    AccessoryBuilder as Accessory,
+} from './AccessoryBuilder'
+
+export {
+    ButtonBuilder as Button,
+} from './ButtonBuilder'
+
+export {
+    ActionRowBuilder as ActionRow,
+} from './ActionRowBuilder'
+
+export {
+    ContainerBuilder as Container,
+} from './ContainerBuilder'