+++ src/event/MessageCommandEvent.ts
@@ -1,4 +1,4 @@
-import type { APIMessage, APIMessageApplicationCommandInteraction } from "..";
+import type { APIMessage, APIMessageApplicationCommandDMInteraction, APIMessageApplicationCommandGuildInteraction, APIMessageApplicationCommandInteraction } from "..";
 import type { Discord } from "../..";
 import { CommandBaseEvent } from "./CommandBaseEvent";
 
@@ -7,5 +7,13 @@ export class MessageCommandEvent<I extends APIMessageApplicationCommandInteracti
     constructor(client: Discord.Client, interaction: I) {
         super(client, interaction);
         this.message = Object.entries(interaction.data.resolved.messages)[0]![1]!
+    }
+
+    override inGuild(): this is MessageCommandEvent<APIMessageApplicationCommandGuildInteraction> {
+        return !!this.interaction.guild
+    }
+
+    override inDM(): this is MessageCommandEvent<APIMessageApplicationCommandDMInteraction> {
+        return !!this.interaction.user
     }
 }