+++ src/event/ModalComponentEvent.ts
@@ -1,4 +1,4 @@
-import { type APIModalSubmitInteraction, ComponentType } from "discord-api-types/payloads";
+import { type APIModalSubmitDMInteraction, type APIModalSubmitGuildInteraction, type APIModalSubmitInteraction, ComponentType } from "discord-api-types/payloads";
 import { ModalBuilder, type MentionableMap } from "../builder/ModalBuilder";
 import type { PathResolver, Prettify } from "../lib/utils";
 import { Discord } from "../..";
@@ -59,5 +59,13 @@ export class ModalComponentEvent<M extends ModalBuilder = ModalBuilder, Path ext
                     }
                 }
             })
+    }
+
+    override inGuild(): this is ModalComponentEvent<M, Path, APIModalSubmitGuildInteraction> {
+        return !!this.interaction.guild
+    }
+
+    override inDM(): this is ModalComponentEvent<M, Path, APIModalSubmitDMInteraction> {
+        return !!this.interaction.user
     }
 }