+++ src/event/UserCommandEvent.ts
@@ -1,4 +1,4 @@
-import type { APIUser, APIUserApplicationCommandInteraction } from "discord-api-types/payloads";
+import type { APIUser, APIUserApplicationCommandDMInteraction, APIUserApplicationCommandGuildInteraction, APIUserApplicationCommandInteraction } from "discord-api-types/payloads";
 import { Discord } from "../..";
 import { CommandBaseEvent } from "./CommandBaseEvent";
 
@@ -7,5 +7,13 @@ export class UserCommandEvent<I extends APIUserApplicationCommandInteraction = A
     constructor(client: Discord.Client, interaction: I) {
         super(client, interaction);
         this.user = Object.entries(interaction.data.resolved.users)[0]![1]!
+    }
+
+    override inGuild(): this is UserCommandEvent<APIUserApplicationCommandGuildInteraction> {
+        return !!this.interaction.guild
+    }
+
+    override inDM(): this is UserCommandEvent<APIUserApplicationCommandDMInteraction> {
+        return !!this.interaction.user
     }
 }